+++ stage-6-backend-code-with-api-routes-fs-react/src/pages/api/feedback/[feedbackId].js
@@ -1,0 +1,14 @@
+import { buildFeedbackPath, extractFeedback } from ".";
+
+function handler(req, res) {
+  const feedbackId = req.query.feedbackId;
+
+  const filePath = buildFeedbackPath();
+  const data = extractFeedback(filePath);
+
+  const selectedFeedback = data.find((feedback) => feedback.id === feedbackId);
+
+  res.status(200).json({ feedback: selectedFeedback });
+}
+
+export default handler;
